fix:【E】経度・緯度情報が空でも保存できるように修正。カラムが変更できるように、doubleからfloatに型を変更

+++ database/migrations/2022_02_11_105656_add_latitude_other1_to_cities_table.php
@@ -7,8 +7,8 @@ use Illuminate\Support\Facades\Schema;
 class AddLatitudeOther1ToCitiesTable extends Migration {
     public function up() {
         Schema::table('cities', function (Blueprint $table) {
-            $table->double('latitude', 4, 2);
-            $table->double('longitude', 5, 2);
+            $table->float('latitude', 4, 2);
+            $table->float('longitude', 5, 2);
         });
     }
